Colocate Registry tests and stabilize CI lint
