+++ v2/identity/src/__tests__/authenticated-http.types-check.ts
@@ -1,0 +1,92 @@
+/**
+ * AuthenticatedHttp keeps its complete signing and verification inputs and
+ * Effect channels exact. These canaries prevent adapters from leaking private
+ * requirements or silently widening the failures consumers must handle.
+ */
+
+import type { HttpClientRequest, HttpServerRequest } from "@effect/platform";
+import type {
+  AgentId,
+  AgentSigningAuthority,
+  AgentSigningError,
+  AuthenticatedHttp,
+  AuthenticationFailedError,
+  MalformedRequestError,
+  OverloadedError,
+  Registry,
+  UnavailableError,
+  VerifiedAgentRequest,
+  VersionMismatchError,
+} from "../index.js";
+import type { Effect, Layer } from "effect";
+
+type Equal<Left, Right> = [Left, Right] extends [Right, Left] ? true : false;
+type Expect<Value extends true> = Value;
+
+type AuthenticationFailure =
+  | MalformedRequestError
+  | AuthenticationFailedError
+  | VersionMismatchError
+  | OverloadedError
+  | UnavailableError;
+
+type SignCall = Readonly<{
+  httpRequest: HttpClientRequest.HttpClientRequest;
+  callerAgentId: AgentId;
+  encodedRequest: unknown;
+  signingAuthority: AgentSigningAuthority;
+}>;
+type VerifyCall = Readonly<{
+  httpRequest: HttpServerRequest.HttpServerRequest;
+  bodyBytes: Uint8Array;
+}>;
+
+type SignEffect = ReturnType<typeof AuthenticatedHttp.signAgentRequest>;
+type VerifyEffect = ReturnType<typeof AuthenticatedHttp.verifyAgentRequest>;
+type AuthenticationLayer = ReturnType<typeof AuthenticatedHttp.layer>;
+
+type SignInputIsExact = Expect<
+  Equal<Parameters<typeof AuthenticatedHttp.signAgentRequest>[0], SignCall>
+>;
+type SignChannelsAreExact = Expect<
+  Equal<
+    SignEffect,
+    Effect.Effect<HttpClientRequest.HttpClientRequest, AgentSigningError>
+  >
+>;
+type VerifyInputIsExact = Expect<
+  Equal<Parameters<typeof AuthenticatedHttp.verifyAgentRequest>[0], VerifyCall>
+>;
+type VerifyChannelsAreExact = Expect<
+  Equal<
+    VerifyEffect,
+    Effect.Effect<
+      VerifiedAgentRequest,
+      AuthenticationFailure,
+      AuthenticatedHttp
+    >
+  >
+>;
+type AuthenticationLayerInputIsExact = Expect<
+  Equal<
+    Parameters<typeof AuthenticatedHttp.layer>[0],
+    Readonly<{
+      liveNonceCapacity: number;
+      agentCardCacheCapacity: number;
+      registryLookupConcurrencyLimit: number;
+    }>
+  >
+>;
+type AuthenticationLayerIsExact = Expect<
+  Equal<AuthenticationLayer, Layer.Layer<AuthenticatedHttp, never, Registry>>
+>;
+
+/** Compile-time evidence for the public AuthenticatedHttp capability. */
+export type AuthenticatedHttpCapabilityCanaries = [
+  SignInputIsExact,
+  SignChannelsAreExact,
+  VerifyInputIsExact,
+  VerifyChannelsAreExact,
+  AuthenticationLayerInputIsExact,
+  AuthenticationLayerIsExact,
+];
