+++ v2/identity/src/registry/__tests__/registry-postgresql.integration.test.ts
@@ -16,7 +16,7 @@ import {
   MOLTZAP_VERSION,
   Registry,
   type VerifiedAgentCard,
-} from "../index.js";
+} from "../../index.js";
 import {
   ADMISSION_CREDENTIAL,
   LOOPBACK_HOST,
@@ -33,9 +33,9 @@ import {
   stopProcess,
   waitForHealthStatus,
 } from "./registry-process.js";
-import { AuthenticationFailedError } from "../http-errors.js";
-import { verifyBootstrapRegistration } from "../registry/admission.js";
-import { makeRegistryStorage } from "../registry/storage.js";
+import { AuthenticationFailedError } from "../../http-errors.js";
+import { verifyBootstrapRegistration } from "../admission.js";
+import { makeRegistryStorage } from "../storage.js";
 import { join } from "node:path";
 import {
   Clock,
