+++ v2/identity/tsconfig.json
@@ -12,7 +12,7 @@
     "src/**/*.test.ts",
     "src/**/*.spec.ts",
     "src/**/*.types-check.ts",
-    "src/__tests__",
+    "src/**/__tests__",
     "dist"
   ],
   "references": []
